gdal_geojson_project/requirements.txt
aspose-cad

C:\Users\yagelm\Desktop\tasks\Gdal\GDAL_proj\archive\GDAL-3.4.3-cp311-cp311-win_amd64.whl